+++ lib/ReactViews/Workbench/Controls/ViewingControls.jsx
@@ -13,6 +13,9 @@ import when from 'terriajs-cesium/Source/ThirdParty/when';
 import classNames from 'classnames';
 import Styles from './viewing-controls.scss';
 
+import createCatalogMemberFromType from '../../../Models/createCatalogMemberFromType';
+import ImagerySplitDirection from 'terriajs-cesium/Source/Scene/ImagerySplitDirection';
+
 const ViewingControls = createReactClass({
     displayName: 'ViewingControls',
     mixins: [ObserveModelMixin],
@@ -55,6 +58,22 @@ const ViewingControls = createReactClass({
         }, 50);
     },
 
+    splitItem() {
+        const item = this.props.item;
+        item.splitDirection = ImagerySplitDirection.LEFT;
+        const serializedItem = item.serializeToJson();
+        serializedItem.name = serializedItem.name + ' (copy)';
+        serializedItem.splitDirection = ImagerySplitDirection.RIGHT;
+        delete serializedItem.id;
+
+        const newItem = createCatalogMemberFromType(item.type, item.terria);
+        newItem.updateFromJson(serializedItem);
+        item.duplicateItemUniqueId = newItem.uniqueId;
+        newItem.duplicateItemUniqueId = item.uniqueId;
+
+        item.terria.showSplitter = true;
+    },
+
     previewItem() {
         let item = this.props.item;
         // If this is a chartable item opened from another catalog item, get the info of the original item.
@@ -69,18 +88,24 @@ const ViewingControls = createReactClass({
 
     render() {
         const item = this.props.item;
+        const canZoom = item.isMappable || (item.tableStructure && item.tableStructure.sourceFeature);
+        const canSplit = defined(item.splitDirection) && item.terria.currentViewer.canShowSplitter && defined(item.clock) && !defined(item.duplicateItemUniqueId);
+        const classList = {[Styles.noZoom]: !canZoom, [Styles.noSplit]: !canSplit, [Styles.noInfo]: !item.showsInfo};
         return (
-            <ul className={classNames(Styles.control, {[Styles.hasZoom]: item.isMappable || item.tableStructure && item.tableStructure.sourceFeature})}>
+            <ul className={Styles.control}>
                 <If condition={item.isMappable}>
-                    <li className={Styles.zoom}><button type='button' onClick={this.zoomTo} title="Zoom to data" className={Styles.btn}>Zoom To Extent</button></li>
+                    <li className={classNames(Styles.zoom, classList)}><button type='button' onClick={this.zoomTo} title="Zoom to data" className={Styles.btn}>Zoom To Extent</button></li>
                 </If>
                 <If condition={item.tableStructure && item.tableStructure.sourceFeature}>
-                    <li className={Styles.zoom}><button type='button' onClick={this.openFeature} title="Zoom to data" className={Styles.btn}>Zoom To</button></li>
+                    <li className={classNames(Styles.zoom, classList)}><button type='button' onClick={this.openFeature} title="Zoom to data" className={Styles.btn}>Zoom To</button></li>
                 </If>
                 <If condition={item.showsInfo}>
-                    <li className={Styles.info}><button type='button' onClick={this.previewItem} className={Styles.btn} title='info'>About This Data Set</button></li>
+                    <li className={classNames(Styles.info, classList)}><button type='button' onClick={this.previewItem} className={Styles.btn} title='info'>About This Data</button></li>
                 </If>
-                <li className={Styles.remove}>
+                <If condition={canSplit}>
+                    <li className={classNames(Styles.split, classList)}><button type='button' onClick={this.splitItem} title="Duplicate and show splitter" className={Styles.btn}>Split</button></li>
+                </If>
+                <li className={classNames(Styles.remove, classList)}>
                     <button type='button' onClick={this.removeFromMap} title="Remove this data" className={Styles.btn}>
                         Remove <Icon glyph={Icon.GLYPHS.remove}/>
                     </button>